Streamline Querying by minimizing number of calls
The edits I made creates a text file for each of the ‘items’ that needs
to be retrieved from Adobe. Instead of conducting a query for report,
elements, metrics and segment every time a report is run, the code
generates a one-time file of all the essential data, and then relies on
the local file to retrieve the information from versus the database.
The information is renewed daily through a file find request.