Try/except for Null names
There are some cases when the data contains a null name value, and so
when one types ```report.data``` to output the result, it'll raise an
error.

I fixed this problem by adding a try/except statement to handle this
error without breaking it. The null name is replaced with the string
value null

```python
            else:
                try:
                    data[element] = row['name'].encode('utf-8')
                except:
                    data[element] = "null"
```